submission/requirements.txt
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
scikit-learn>=1.2.0
scipy>=1.10.0
